fix installation_issues

+++ ompi/mpi/c/pgcopy.h
@@ -4,7 +4,7 @@ void intToBinary(int integer, char* buffer, int* offset);
 void doubleToBinary(double value, char* buffer, int* offset);
 void stringToBinary(char* string, char* buffer, int* offset);
 void timestampToBinary(struct timespec time, char* buffer, int* offset, int round_seconds);
-void byteaToBinary(int* array, int length, char* buffer, int* offset, int p2p);
+void byteaToBinary(uint8_t* array, int length, char* buffer, int* offset, int p2p);
 void newRow(char* buffer, int column_count, int* offset);
 //void createTrailer(char* buffer, int* offset);
